Remove unnecessary offers state variable
Reason: it was supposed to be used for optimisation - to avoid iteration through the whole deliveries table. However without implementing an algorithm for removing an item in the middle of a table, instead of removing an id from offers table we would set it to 0. Effectively the offers table would have the same length as deliveries table.